+++ common/cfm2.c
@@ -13,7 +13,7 @@ FILE* lookInRoot(char* filename, char* type){
 
 // returns the CL folder path in InCGames parent folder
 char* getRootFilePath(){
-    return "C:\\InCGames\\cl\\";
+    return "C:/InCGames/cl/";
 }
 
 WIN32_FIND_DATA* listFiles(char* key, char* path){
@@ -25,10 +25,10 @@ WIN32_FIND_DATA* listFiles(char* key, char* path){
 
     if(!path){
         snprintf(fullPath, sizeof(fullPath), "%s*.*", getRootFilePath());
-    }else if(strstr(path, "C:\\") == NULL){
-        snprintf(fullPath, sizeof(fullPath), "%s%s\\*.*", getRootFilePath(), path);
-    }else if(strstr(path, "\\*.*") == NULL){
-        snprintf(fullPath, sizeof(fullPath), "%s\\*.*", path);
+    }else if(strstr(path, "C:/") == NULL){
+        snprintf(fullPath, sizeof(fullPath), "%s%s/*.*", getRootFilePath(), path);
+    }else if(strstr(path, "/*.*") == NULL){
+        snprintf(fullPath, sizeof(fullPath), "%s/*.*", path);
     }
     printf("[DEBUG] filepath: %s\n", fullPath);
     hfind = FindFirstFileA(fullPath, &wfd);
@@ -58,7 +58,7 @@ char* getListedDirectories(char* directory){
     char path[1024];
     strncpy(path, directory, sizeof(path)-4);
     path[sizeof(path)-1] = '\0';
-    strcat(path, "\\*");
+    strcat(path, "/*");
     
     WIN32_FIND_DATAA findData;
     HANDLE hFind = FindFirstFileA(path, &findData);