+++ projects/InLine/errorHandler.c
@@ -1,0 +1,5 @@
+#include "errorHandler.h"
+
+void logError(inlineErrorCode* code){
+
+}