+++ projects/InLine/inLine.c
@@ -3,6 +3,7 @@
 void textStart(Verb* devVerbs){
     
 }
+
 void textUpdate(char* text){
     text = textLogger();